add exists check on backend test teardown method

+++ Backend/simplicityserver/test/server_routes_test_case.py
@@ -49,8 +49,10 @@ class ServerRoutesTestCase(unittest.TestCase):
 	@classmethod
 	def tearDownClass(cls):
 		cls._conn.close()
-		os.remove(TESTDBPATH)
-		os.remove(TEMPTHUMBPATH)
+		if os.path.exists(TESTDBPATH):
+			os.remove(TESTDBPATH)
+		if os.path.exists(TEMPTHUMBPATH):
+			os.remove(TEMPTHUMBPATH)
 
 	def tearDown(self):
 		cur = self._conn.cursor()
